Move the config file to a subdirectory
This is because we're about to create an additional config file, so it
would be nice to have them in the same place.
